removed some extra files, renamed main to snipebot so i knew what it was and also cleaned up a little bit of it
